% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{Gene set annotations}
\format{
A data frame with 243 rows and 17 variables:
\describe{
  \item{GeneSetModule}{gene set names}
  \item{Annotation_}{annotations}
  \item{Description}{gene set descriptions}
  \item{Link}{link to raw data}
}
}
\source{
see \code{Link} columns
}
\usage{
annotations
}
\description{
Annotations are structured as a column of gene set names followed by a number
of annotation columns containing annotations for that respective gene set.
Column names are optional, there is no limit on the number of annotation
columns, and additional columns are allowed.
}
\keyword{datasets}
