% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{data}
\name{database}
\alias{database}
\title{Gene set database}
\format{
A list containing two elements:
\describe{
  \item{gs_genes}{a list of gene sets and the genes they contain}
  \item{gs_info}{a data frame of gene set metadata}
}
}
\source{
see \code{\link{annotations}}
}
\usage{
database
}
\description{
Gene set database
}
\keyword{datasets}
