% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{Gene set annotations}
\format{
A data frame with 243 rows and 16 variables:
\itemize{
  \item{\code{name}} gene set names
  \item{\code{anno_x}} annotation column x
  \item{\code{info}} gene set origin
}
}
\source{
see \code{info} columns
}
\usage{
annotations
}
\description{
Gene set annotations
}
\keyword{datasets}
