% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{data}
\name{database}
\alias{database}
\title{Gene set database}
\format{
A data frame with 243 rows and 6780 variables:
\describe{
  \item{X1}{gene set names}
  \item{X_}{genes}
}
}
\source{
@seealso \code{\link{annotations}}
}
\usage{
database
}
\description{
Databases are structured as a column of gene set names followed by a number
of gene columns containing genes for that respective gene set.
Column names are optional, there is no limit on the number of gene columns,
and additional columns are allowed.
}
\keyword{datasets}
